Displays memory usage statistics after build
Given the limited memory of most Arduino boards, knowing a sketch's
memory usage is important to developers. The Arduino IDE displays this
and provides warnings when memory usage is high. The IDE includes a
tool (avr-size) that calculates memory usage.

This patch runs the tool and compares memory (both text and data) to the
maximums as stored in the boards.txt file.